python/task-management-system/loanhub/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
alembic
psycopg2-binary
pytest
httpx
python-multipart
python-jose[cryptography]
passlib[bcrypt]
